# swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "NeuroSuitUp API",
        "description": "Tools for communicating with the NeuroSuitUp database",
        "version": "1.0.0",
        "contact": {
        "name": "Apostolos Praftsiotis",
        "email": "apospraf@gmail.com"
        }
    },
    "servers": [
        {
            "url":"http://localhost:3000",
            "description":"Temp localhost server"
        },
        {
            "url": "https://neurosuituptest.herokuapp.com/",
            "description": "Heroku free server"
        }, 
        {
            "url": "http://192.168.0.188:3000",
            "description": "Heroku free server"
        }
    ],
    "components": {
      "schemas": { 
...
        "ExerciseInsert": {
            "type": "object",
            "properties": {
                "codename": {
                    "type": "string",
                    "description": "The codename of the specific exercise"
                },
                "isPublic": {
                    "type": "boolean",
                    "description": "Whether or not the specific exercise is available to every instructor"
                },
                "tasks": {
                    "type": "array",
                    "description": "Array containing the tasks the define the specifc exercise",
                    "items":{
                        "type": "string",
                        "description": "The task id (ObjectId, mongodb)"
                    }
                },
                "delays": {
                    "type": "array",
                    "description": "The delay in seconds for each task pair",
                    "items": {
                        "type": "string",
                        "description": "The delay in seconds (sec)"
                    }
                }
            }
        },
        "Exercise": {
            "allOf": [
                {"$ref": "#/components/schemas/ExerciseInsert"},
                {
                    "type": "object",
                    "properties": {
                       "_id": {
                           "type": "string",
                           "description": "The exercise id (ObjectId, mongodb)"
                        }
                    }
                }
            ]
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }, 
        "Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
      },
      "parameters": {
          "selectstring": {
              "in": "header",
              "name": "selectstring",
              "description": "Selects the fields of the returned resource",
              "schema": {
                  "type": "string"
              },
              "allowEmptyValue": true
          }
      },
      "responses": {
...
        "Unauthorized": {
            "description": "Authorization failed, either by unsuccessfull verification of token, or by limited permissions",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error"
                    }
                }
            }
        }, 
        "InternalError": {
            "description": "Internal or structural error",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error"
                    }
                }
            }
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
...
      "/exercises": {
        "get": {
            "summary": "Get all available exercises for the logged in instructor",
            "tags": ["exercises"],
            "parameters": [{
                "$ref": "#/components/parameters/selectstring"
            }], 
            "responses": {
                "200": {
                    "description": "Query successfull, return all found exercises",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "count": {
                                        "type": "integer"
                                    },
                                    "tasks": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Exercise"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        },
        "post": {
            "summary": "Insert a new exercise to the database",
            "tags": ["exercises"],
            "requestBody": {
                "description": "The new exercise",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "$ref": "#/components/schemas/ExerciseInsert"
                        }
                    }
                }
            },
            "responses": {
                "201": {
                    "description": "Exercise successfully inserted to the database",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "409": {
                    "description": "Exercise codename already exists",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }  
    },
    "/exercises/{exerciseId}": {
        "parameters": [{
            "in": "path",
            "name": "exerciseId",
            "description": "The requested exercise id",
            "required": true,
            "schema": {
                "type": "string"
            }
        }],
        "get": {
            "summary": "Get exercise by id",
            "tags": ["exercises"],
            "parameters": [{
                "$ref": "#/components/parameters/selectstring"
            }],
            "responses": {
                "200": {
                    "description": "Query was successful, return found exercise",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "task": {
                                        "$ref": "#/components/schemas/Exercise"
                                    }
                                }
                            }
                        }
                    }
                }, 
                "404": {
                    "description": "Exercise was not found",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        },
        "delete": {
            "summary": "Delete exercise by id",
            "tags": ["exercises"],
            "responses": {
                "200": {
                    "description": "Successfully deleted exercise",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "403": {
                    "description": "Exercise is not owned, thus cannot be deleted",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "409": {
                    "description": "Deletion conflict",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Deleting this exercise will affect the prescriptions of some users"
                                    },
                                    "users": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "description": "The username of the affected user"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
    },
    "/fetchExercises": {
        "post": {
            "summary": "Fetch specific exercises with ids, located at the body of the request",
            "tags": ["exercises"],
            "parameters": [
                {
                    "$ref": "#/components/parameters/selectstring"
                }
            ],
            "requestBody": {
                "description": "The array of the requested exercise ids",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "description": "The exercise id (ObjectId mongo)"
                            }
                        },
                        "example": "{['615c94503b3ffd5cccec496a', '615c945e3b3ffd5cccec496b']}"
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Successfully fetched all exercises",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "count": {
                                        "type": "integer",
                                        "description": "The number of found exercises"
                                    },
                                    "tasks": {
                                        "type": "array",
                                        "description": "Array containing the exercise ids",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
    }
    }
}